[stubgen] Rename x86_64 architecture to x86 to mesh with Trusty LK build system
stubgen.py is used to generate syscall implementation, in order to
generate x86 64-bit syscall implementation, x86_64 was introduced in
this script. Argument --arch needs to be sepcified as x86_64 when
invoking this script to generate x86 64-bit implementation.

Currently, this script is invoked by makefile to generate syscall
implementation automatically, and argument --arch is passed with
marco ARCH defined in Trusty LK. Both x86 and x86 64-bit define ARCH as
x86, and use SUBARCH x86-32 and x86-64 to distinguish. Definitions are
incompatible to design of this script and Android.

Meanwhile Trusty supports x86 64-bit only, update script to generate x86
64-bit syscall implementation when argument --arch is passed with x86.

Bug: 119111590
Change-Id: Ia1e2882b8493a653d6a8c7e2f912cf2ebb1920ce