[lib][trusty][ipc] Skip closed connections during startup port shutdown
Skipping these connections avoids restarting the app for a connection
that has already been shutdown.

Also, shutdown a client channel before checking if it is in the
waiting_for_port list. This avoids a race where a closed channel might be
added to the list by a port that is shutting down before its state is
marked as closed but after the client has checked for it in the
waiting_for_port list.

Bug: 121390945
Change-Id: If3fbe4d7adc1163df29374ad4ae541d6980a4358