[lib][memlog] Do not expose memlog data until the writer is done.
Due to how LK's printf works, a single print statement may result in
multiple write calls to the memlog. A second thread reading the memlog
can observe each write as it occurs. This can result in corrupted output
if the observer treats each write as a separate logging event.

To work around this issue, a seperate "commit" function was added to
signal that a write is complete and that memlog should expose the
recently written data.

Bug: 122574827
Change-Id: I5db21f774bae4a6445793460ee934551a6eab2e6