[appmgmt] Refactor the handling of startup ports
Refactor the handling of startup ports in preparation for [un]loadable
apps.

Startup ports are now tracked as part of the app properties and the
interactions between ipc and appmgr now use port names. This eliminates
the need to keep to separate tracking structures in sync (port list and
app list).

Bug: 115420908
Change-Id: I318e1aac7025f3c5794692fad3c20fa7e31db8c4