[appmgmt] Refactor locking in preparation for [un]loadable applications
Accesses to the application list while an application could be removed
should be locked.

Don't expose trusty_app_find_by_uuid as it returns a raw pointer to an
app that could be invalidated at any time.

Add _locked suffix to helper functions that must be called with the
app_lock held.

Bug: 115420908
Change-Id: I178e283cf8add45cbe41ac8575f61495d6725c29