[appmgmt] Add application management syscalls
sys_register_app can be used to add applications to be managed by the
application manager.

There is now a distiction between apps that are present in the kernel
image (built-in) and apps that are created by using the register
syscall. Registered applications are started at registration time unless
they have the DEFERRED_START manifest flag.

sys_unregister_app can be used to remove applications managed by the
application manager that were previously added with sys_register_app.

An application will only be fully removed when it is in the NOT_RUNNING
state. If sys_unregister_app is called on an application that is in any
other state at that time, the application will be flagged for removal
and will be removed the next time it exits.

These syscalls are only available in test builds

Bug: 115420908
Change-Id: I523b3f2b19958719d9266b304dad5ff5ff73828e