[app][busytest] Add server to make every CPU busy
Add a server that makes spins on every CPU when a client connects to
com.android.kernel.busy-test. When the connection is closed it goes back
to idle. This can be used from other tests, or in combination with other
tests, to thest the system when CPUs don't go idle.

Bug: 129157247
Change-Id: Iad7bb5f587e86d5d864a7bea96cee84d03610061