Add lower bound check to handle_id
This check is technically unnecessary, as in the event of a "negative"
idx, we will get a very high idx, which will be lower than
IPC_MAX_HANDLES. However, raising IPC_MAX_HANDLES sufficiently high
or making idx an int would turn this into a bug. As it is, it also trips
UBSan.

Adding this check makes our code more obviously correct, and hardens
against future modifications.

Bug: 129300035
Change-Id: I382514d68a1ca57f4f4913f0888c5f7b91d057bc