reflist: Add obj_has_ref and check obj_add_ref
Adds an API to check whether at least one outstanding reference exists
to an object.

Additionally, check when adding a ref that there is already at least one
outstanding ref, (since initialization creates one, and the object
should be destroyed at zero).

Additionally, this change adds obj_add_ref_allow_unreferenced_obj which
omits this check for case like the storage block cache which wishes to
reuse entries which have dropped to zero references, and for which this
an expected condition.

Bug: 117221195
Change-Id: I1aab8a51f1de600abbed89369620ce5255042ace