Add common kernel libs to userspace library build system
libc-ext and ubsan are used in both the kernel and by userspace
libraries. Since we want to use them in userspace, we need to switch to
building with the library.mk system instead of as a generic lk module.
The variables specific to userspace are ignored if we are building the
module for the kernel.

Test: build.py qemu-generic-arm64-test-debug
Bug: 169448829
Change-Id: Icdb5cfa1b6a385eb990e580f04e527447a6c29f9