[lib][sm] Allow SMC_FC_FIQ_ENTER after panic on core that called panic
If an fiq triggers after platform_halted allow the SMC_FC_FIQ_ENTER call
to go through to the gic driver. Otherwise we would try to return to the
non-secure world with the fiq still pending and it would immedeatly call
SMC_FC_FIQ_ENTER again.

This was already fixed for the cores that did not trigger the panic.
Call the sm_sched_nonsecure_fiq_loop on the core that called panic
instead of directly calling sm_sched_nonsecure so that it can dispatch
the SMC_FC_FIQ_ENTER call the same way as the other cores.

Bug: 186042636
Change-Id: I4e6bf6a11d9d8708f12cda08319b1ce892c0f538