[lib][ubsan] Remove UBSAN_HARD_FAIL option
This option is never intentionally set to false. However, it can cause
complications if CFI is enabled, but UBSan is not. In this case the flag
needs to be explicitly set to true. It simpler to just avoid the
complexity.

Also, the behavior of UBSAN_HARD_FAIL := false is ill-defined, since
there is no way to guarantee a recovery from a UB.

Bug: 190293335
Test: trigger UBSan failure
Change-Id: Ia05586780c3b45231ebeeed1ce791e77a93526db