[app][userscstest] Fix user shadow call stack test
When shadow call stacks are enabled for libc and for a trusty app, the
expecation that the first element on the shadow call stack is non-zero
is incorrect. The link register is initialized to zero so we instead
expect that the second shadow call stack slot is non-zero for all apps.

Bug: 79994372
Change-Id: Iad6f5223fbfb20f3c3727e73323ccf49abb97659