Use cached trusty app paths rather than re-creating them
The userspace build used to re-create paths to the app ELF files when it
needed them. This meant that we could not change app ELF paths or module
names without coordinating that across the build system. Now that each
app caches the path to its output file, we can use these cached values
instead of re-creating the paths.

Test: build.py
Bug: None
Change-Id: If9bfb1bcdbad12b3d747bb757e7b3aede4841f5a