[lib][trusty] Return NO_ERROR when app has already started
trusty_app_create_and_start() creates and starts an
application in two steps, and drops the apps_lock between
the two. There is a possible race where an IPC connection
comes in and starts the application before the second step,
causing request_app_start_locked() to return ERR_ALREADY_STARTED.
This would previously be returned to apploader which would
interpret it as a failure to load the application.
This CL changes the return code to NO_ERROR so the apploader
considers the application as loaded.

Bug: 205782087
Change-Id: Ic92a057b5e3a00714d8af57fba2c7475696a5543