[lib][backtrace] Do not print absolute kernel addresses
For security reasons, we do not want to print randomized
kernel pointers in stack traces in release builds.
This patch subtracts the dynamic load bias so that all
printed kernel addresses match the canonical ones
from the kernel ELF file for release builds, and prints
both the actual and the canonical addresses for debug builds.

Test: com.android.trusty.crashtest
Bug: 218504830
Change-Id: I95b234cf8d4bfa7adc728f14de4e6fcaaff5b011