lib/libc-trusty: Add rules.mk for musl in kernel
This adds a rules.mk for musl in the kernel in lib/libc-trusty.
This new module also adds writev, functions for mapping files
to io_handle_t with musl and modifies the modules that do this
to use the new functions (aosp/2269512).

This commit also adds a way to store thread-local libc state
as a TLS entry. This libc_state struct is analogous to the
pthread_t used in musl's arch-specific pthread_arch.h
headers. Userspace is currently single-threaded so most
pthread code is stubbed out without issue, but the kernel is
multithreaded so relying on stubbed out pthread code may
lead to issues that can assign invalid values to pthread_t's
fields. While we could implement the pthreads API on top of
LK's thread API, for now we just define our own small struct
for thread-local info (errno and locale for stdio) and stick
it in a TLS entry (only when building with musl).

Bug: 230134581
Change-Id: I0dfdbbb0e2258692131888db9794e13c2e6e2986