app: userscstest: Fix outdated app state check
A recent change (aosp/2132908) introduces additional states for
trusty apps. In some of these states, the thread field of the
trusty_app structure may be NULL. Update condition to skip checks
of shadow stacks for trusty apps that are not running and thus
may not have a thread. To ensure that the checks do not pass in
cases where there are no running apps, code is added to check
that at least one trusty app was running when shadow stacks were
inspected.

Test: build.py slider-test-debug --test userscstest
Bug: 238856863
Change-Id: I1db0af9390d8630651b5816aa5e7b30002e562d7