lib/mmutest: Replace ADR with ADRL in mmutest for ARM64
Building with libbinder in the kernel seems to increase the
size of the .text section just enough that it pushes the
label (in .data) in an ADR instruction in mmutest out of
range. This causes a linker error about the relocation for
this instruction being out of range. This change replaces it
with the equivalent ADRL pseudo-instruction to allow
referencing the label.

Bug: 230134581
Change-Id: Ief460b70f1079c1ea62a75067925f422b4cedbb7