libc-trusty: Use lk stdio instead of musl's
Use most of the stdio code from lk instead of musl.
printf in the kernel can be called from both interrupt
and thread context. The lk console code handles this
case correctly with separate locks, so we revert to it
to avoid problems caused by musl not handling this case.

Bug: 230134581
Change-Id: Ied7d5d7c5c6fc51ceb4cb6c3acc4dde07cbb593f