user-tasks.mk: Override including tests in image
Allow targets to configure which tests should be built into the trusty
image itself rather than built only as loadable apps. Defaults to
including all tests in the image and building them as loadable apps as
well. This can be overridden with the TRUSTY_BUILTIN_USER_TESTS
environment variable to select a particular subset of tests to include
in the image.

Test: build.py generic-arm64-test-debug
Bug: 263426706
Change-Id: I5d4ad322bb8acefa4e8aa58fe26983ec47834ed9