make/dtb.mk: Generate .c with INCBIN for .dtb
This commit changes dtb.mk to generate the .c with the
.incbin directive for embedding a dtb as part of the build
process. This simplifies the usage of dtb.mk by only
requiring users to specify a .dts. This change makes the
symbol for each dtb the path to the module (backslashes
replaced by underscores) though accessing embedded dtbs with
the dtb_embedded library is still preferred.

Bug: None
Change-Id: Ia5b08ed73fa956c19bde34c59cf6a4327c8f6fd4