user-tasks.mk: Only build loadable apps if the apploader service is included
Building all tests as loadable apps by default does not work if the
target build does not support loadable apps. We should not build any
loadable apps for build configurations that do not include the apploader
service, as there is no way to run these apps. This change disables
loadable app building if the apploader is not included, and exits with a
build error if loadable apps were explicitly requested without the
apploader.

Test: build.py
Bug: 296257715
Change-Id: I6486718619c0b2ba8b720866faa28557f69b0363